+++ test/bench/opt/powell4_function_test.cpp
@@ -26,6 +26,7 @@
 #include "num_collect/opt/dfp_optimizer.h"
 #include "num_collect/opt/dividing_rectangles.h"
 #include "num_collect/opt/downhill_simplex.h"
+#include "num_collect/opt/heuristic_global_optimizer.h"
 #include "num_collect/opt/steepest_descent.h"
 
 STAT_BENCH_MAIN
@@ -119,6 +120,18 @@ STAT_BENCH_CASE_F(
     powell4_function_fixture, "opt_powell4_function", "dividing_rectangles") {
     STAT_BENCH_MEASURE() {
         auto optimizer = num_collect::opt::dividing_rectangles<
+            num_prob_collect::opt::powell4_function>();
+        const auto [lower, upper] = search_region();
+        optimizer.init(lower, upper);
+        this->test_optimizer(optimizer);
+    };
+}
+
+// NOLINTNEXTLINE
+STAT_BENCH_CASE_F(powell4_function_fixture, "opt_powell4_function",
+    "heuristic_global_optimizer") {
+    STAT_BENCH_MEASURE() {
+        auto optimizer = num_collect::opt::heuristic_global_optimizer<
             num_prob_collect::opt::powell4_function>();
         const auto [lower, upper] = search_region();
         optimizer.init(lower, upper);