+++ test/bench/opt/quadratic_function_test.cpp
@@ -24,7 +24,7 @@
 
 #include "num_collect/opt/dividing_rectangles.h"
 #include "num_collect/opt/golden_section_search.h"
-#include "num_collect/opt/heuristic_1dim_optimizer.h"
+#include "num_collect/opt/heuristic_global_optimizer.h"
 #include "num_collect/opt/sampling_optimizer.h"
 
 STAT_BENCH_MAIN
@@ -100,9 +100,9 @@ STAT_BENCH_CASE_F(quadratic_function_fixture, "opt_quadratic_function",
 
 // NOLINTNEXTLINE
 STAT_BENCH_CASE_F(quadratic_function_fixture, "opt_quadratic_function",
-    "heuristic_1dim_optimizer") {
+    "heuristic_global_optimizer") {
     STAT_BENCH_MEASURE() {
-        auto optimizer = num_collect::opt::heuristic_1dim_optimizer<
+        auto optimizer = num_collect::opt::heuristic_global_optimizer<
             num_prob_collect::opt::quadratic_function>();
         const auto [lower, upper] = search_region();
         optimizer.init(lower, upper);